requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
apsw==3.36.0.post1
apturl==0.5.2
asgiref==3.5.2
astroid==2.12.13
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
black==21.10b0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cchardet==2.1.7
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
cloudscraper==1.2.60
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
constantly==15.1.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
css-parser==1.0.7
cssbeautifier==1.14.7
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
dill==0.3.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.1
django-ckeditor==6.5.1
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djlint==1.19.5
dnspython==2.1.0
EditorConfig==0.12.3
entrypoints==0.4
execnet==1.9.0
executing==1.0.0
feedparser==6.0.8
filelock==3.6.0
flake8==4.0.1
frozenlist==1.3.1
gpg===1.16.0-unknown
greenlet==1.1.2
h11==0.13.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==0.14.5
httplib2==0.20.2
httpx==0.22.0
hyperlink==21.0.0
idna==3.3
ifaddr==0.1.7
importlib-metadata==5.0.0
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.15.3
ipython==8.5.0
isort==5.6.4
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jsbeautifier==1.14.7
jupyter-core==4.11.1
jupyter_client==7.3.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.8.0
livereload==2.6.3
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mechanize==0.4.7
mkdocs==1.1.2
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.2
mypy==0.942
mypy-extensions==0.4.3
nala==0.11.1
natsort==8.0.2
nest-asyncio==1.5.5
netifaces==0.11.0
nose2==0.9.2
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
praw==7.6.0
prawcore==2.3.0
price-parser==0.3.4
prompt-toolkit==3.0.31
protobuf==3.12.4
protonvpn-cli==3.12.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py7zr==0.11.3+dfsg
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycares==4.2.2
pychm==0.8.6
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.15.9
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-sugar==0.9.4
pytest-xdist==2.5.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-gnupg==0.4.8
pythondialog==3.5.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
pyzmq==24.0.0
regex==2022.10.31
reportlab==3.6.8
repoze.lru==0.7
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==11.2.0
Routes==2.5.1
ruamel.yaml==0.16.13
SecretStorage==3.3.1
selenium==3.141.0
service-identity==18.1.0
sgmllib3k==1.0.0
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.5.0
systemd-python==234
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
Twisted==22.1.0
typed-ast==1.4.3
typer==0.4.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
udemy-enroller==4.1.2
ufw==0.36.1
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==3.7.0
webencodings==0.5.1
websocket-client==1.4.1
wrapt==1.14.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarl==1.8.1
zeroconf==0.38.3
zipp==1.0.0
zope.interface==5.4.0
